requirements.txt
asgiref==3.8.1
boto3==1.37.28
botocore==1.37.28
dj-database-url==2.3.0
Django==5.2
django-cors-headers==4.7.0
django-storages==1.14.6
gunicorn==23.0.0
jmespath==1.0.1
packaging==24.2
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
s3transfer==0.11.4
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.1
urllib3==2.3.0
whitenoise==6.9.0
